refactor(tanstack-table): divide utils into different files
